requirements.txt
openai>=1.30.0
streamlit>=1.40.0
arxiv>=2.1.0
chromadb>=0.5.0
redis>=5.0.0
duckduckgo-search>=6.0.0
python-dotenv>=1.0.0
